Update: changed id type
id type changed from uint8_t to uint16_t as CAN message id is actually
11 bits long. This allows for up to 2047 different ids to be used
	modified:   src/mcp2515_driver.c
	modified:   src/mcp2515_driver.h
	modified:   test/test_mcp2515_driver.c

+++ src/mcp2515_driver.c
@@ -346,7 +346,7 @@ bool mcp2515_rx0_is_full(void)
 
 // TODO Remove all the direct spi calls so that they are kept in the same place
 // It might be better to leave here and just try and improve the efficiency instead
-void mcp2515_driver_read_can_message(uint8_t * id, uint8_t * len,
+void mcp2515_driver_read_can_message(uint16_t * id, uint8_t * len,
                                      uint8_t * read_buf)
 {
     uint8_t buf[14];
